requirements.txt
beautifulsoup4==4.12.2
ffmpeg-python==0.2.0
flask==2.2.3
minio==7.1.14
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-requests==0.39b0
requests==2.28.2
yt-dlp==2023.3.4